requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
jinja2>=3.1.0
python-multipart>=0.0.5
